feat: mingw64 compile

+++ include/uuid.h
@@ -19,7 +19,7 @@ static inline void uuid_generate(char s[37])
 	CoCreateGuid(&guid);
 
 	snprintf(s, 36, "%08x-%04x-%04x-%02x%02x-%02x%02x%02x%02x%02x%02x",
-		guid.Data1, guid.Data2, guid.Data3, (unsigned int)guid.Data4[0], (unsigned int)guid.Data4[1],
+		(unsigned int)guid.Data1, (unsigned int)guid.Data2, (unsigned int)guid.Data3, (unsigned int)guid.Data4[0], (unsigned int)guid.Data4[1],
 		(unsigned int)guid.Data4[2], (unsigned int)guid.Data4[3], (unsigned int)guid.Data4[4],
 		(unsigned int)guid.Data4[5], (unsigned int)guid.Data4[6], (unsigned int)guid.Data4[7]);
 }
